requirements.txt
absl-py
alphafold
#alphafold.common
#alphafold.data
#alphafold.data.tools
#alphafold.model
#alphafold.relax
datetime
google.colab
hashlib
ipywidgets
itertools
json
jsonpickle
matplotlib
numpy
oauth2client
os
pathlib
pickle
py3Dmol
pydrive.auth
pydrive.drive
random
re
requests
#shutil
string
sys
tarfile
tensorflow
time
tqdm
traceback
warnings
zipfile